draft concept for ble-split to transport refactor
